baekjoon

+++ src/level1/level1_10/Main.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        long a = scanner.nextInt(), b = scanner.nextInt();
+        int a = scanner.nextInt(), b = scanner.nextInt();
         System.out.println(a * (b % 10));
         System.out.println(a * ((b / 10) % 10));
         System.out.println(a * ((b / 100) % 10));
