+++ src/level1/level1_4/Main.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        long a, b;
+        int a, b;
         do {
             a = scanner.nextInt();
             b = scanner.nextInt();
