+++ src/level1/level1_8/Main.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        long year;
+        int year;
         do {
             year = scanner.nextInt();
         } while (1_000 > year || year > 3_000);
