+++ src/level2/level2_3/Main.java
@@ -5,9 +5,9 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        long year;
+        int year;
         do {
-            year = scanner.nextLong();
+            year = scanner.nextInt();
         } while (1 > year || year > 4_000);
         if (year % 4 == 0 && (year % 100 != 0 || year % 400 == 0)) {
             System.out.println(1);
